requirements.txt
Django>=4.0
psycopg2
python-dotenv
djangorestframework
djangorestframework-simplejwt
